Support permission_broker for chrome.hid devices.
This enables the Chrome to query permission_broker for access
to hidraw device paths when running on Chrome OS.

HidService has been refactored so that it is owned by the
extensions API embedder, rather than being a singleton.

Because hid has just moved to //extensions as well, this also establishes a new ShellExtensionsAPIClient to create a HidService for app_shell.

BUG=367145
TEST=manual test on ChromeOS dev build
TBR=derat@chromium.org for mechanical app_shell changes

Review URL: https://codereview.chromium.org/419713002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@285774 4ff67af0-8c30-449e-8e8b-ad334ec8d88c