Limit the window of time a promo can be visible after first view.
Notification promos currently have start, end times, and a number of maximum
views. However, the length of time a promo can be visible was not capped.
This change adds a flag to cap the length of time a promo is visible by
max_seconds_.

BUG=None
Test=None

Review URL: https://codereview.chromium.org/419033007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@285803 4ff67af0-8c30-449e-8e8b-ad334ec8d88c