Fix WiFi config view for previously configured client cert.
Before, wifi_config_view expected to read only a plain PKCS#11 ID from the shill::kEapCertIdProperty. However the format changed to '<slot_id>:<key_id>' since the system token is enabled in ChromeOS. This lead to the view not correctly pre-selecting the previously configured certificate.
With this change, the property is correctly parsed and the pre-selection works again for certificates from the user's slot.

BUG=358366, 396181

Review URL: https://codereview.chromium.org/410113002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@285806 4ff67af0-8c30-449e-8e8b-ad334ec8d88c