Fix how we use the new profile management flags.
We are first going to release --new-avatar-menu, which is essentially
what --new-profile-management used to be, minus the lock functionality.
(--new-avatar-menu used to be an opt-in intermediate step, so functionality
needs to be shuffled around.

Basically this is a lot of s/new-profile-management/new-avatar-menu/g.

BUG=391497
TEST=Start Chrome with --new-avatar-menu. You should see the new avatar
menu, but not the lock button on any of the signed in profiles.

Review URL: https://codereview.chromium.org/415643002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@285942 4ff67af0-8c30-449e-8e8b-ad334ec8d88c