Fix timing edge case for OEM folder sync
If the OEM app arrives before we initially receive the AppList sync data (which can happen with a newly created account), the following event chain will occur:

AddItem(oem app) -> FindOrCreateOemFolder() -> model_->AddItem(oem folder) -> ModelObserver::OnAppListItemAdded(oem folder) -> AddOrUpdateFromSyncItem(oem folder)

If AddOrUpdateFromSyncItem creates a sync item, the logic in ResolveFolderPositions will treat the account as an existing account instead of as a new account.

To prevent that, we skip adding a sync item when an OEM folder is added, relying on the code in ResolveFolderPositions to add the sync item instead.

BUG=chrome-os-partner:29764

Review URL: https://codereview.chromium.org/421213012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@286095 4ff67af0-8c30-449e-8e8b-ad334ec8d88c