Replace forward declaration with include in default_component_installer.h
Because the "DictionaryValue" type is used in a scoped_ptr and scoped_ptr needs
to invoke the destructor, the full definition of the class is needed at compile
time. We apparently get lucky with include order on the non-Windows platforms,
because this only showed up when adding a new transitive dependency upon
default_component_installer.h and then only failed on Windows.

Fixing this unblocks the unconditional inclusion of the
cld_component_installer, which we need in order to fix crbug/367239.

BUG=
TBR=waffles@chromium.org

Review URL: https://codereview.chromium.org/422253002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@286200 4ff67af0-8c30-449e-8e8b-ad334ec8d88c