[AiS] Support prefetch query.
Answers needs to signal to search if there is a reasonable guess at the
full query text. For this purpose, it will set pfq= to the full query
text and qha= to the answers type.

A "reasonable guess" is given if the current query text is a strict
prefix of the full query text of an Answer that has been displayed
before, iff that Answer has occupied the top slot/second slot.

Currently, the cache of answers displayed before is one element deep.

R=pkasting@chromium.org
BUG=370192

Review URL: https://codereview.chromium.org/420723007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@286413 4ff67af0-8c30-449e-8e8b-ad334ec8d88c