Add a new "Configure" mechanism to CLD browser-side data interfaces.
This allows test code to be included all the time, regardless of whether a
non-static CLD build is in use. It also decouples sources from needing to
include an implementation-specific header to configure CLD, which eliminates
much of the need for the "#if defined" statements in code. Coupled with
the new CldDataSource (see related bug), this will allow a significant
refactoring of the build logic which will make it possible to build
chrome_shell and other targets with target-specific CLD configurations.

BUG=367239
TBR=toyoshim@chromium.org

Review URL: https://codereview.chromium.org/424053002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@286430 4ff67af0-8c30-449e-8e8b-ad334ec8d88c