Retrieve the list of keyboard layouts for public sessions async
CL 398543002 introduced a GetKeyboardLayoutsForLocale() method that
synchronous resolves the requested locale to an actually available UI
locale and returns the keyboard layouts available for it. However,
resolving a locale requires file I/O, which is not allowed on the UI
thread. This led to DCHECK failures.

This CL makes GetKeyboardLayoutsForLocale() asynchronous. It first
resolves the locale on a background thread and then retrieves the list
of keyboard layouts for it on the UI thread.

BUG=241790
TEST=Debug build no longer DCHECKs when choosing public session locale
R=alemate@chromium.org, dzhioev@chromium.org

Review URL: https://codereview.chromium.org/417103003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@286571 4ff67af0-8c30-449e-8e8b-ad334ec8d88c