Web Notifications should always reappear as popups when updated.
The message center already does this for notifications with a higher
priority than the notification they replace, but this makes it explicit
for Web Notifications (which cannot indicate priority).

This patch re-enables the NotificationTest.TestNotificationReplacement
test, which was marked as DISABLED because of LSan failures.

BUG=366102, 366539

Review URL: https://codereview.chromium.org/301713002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@286831 4ff67af0-8c30-449e-8e8b-ad334ec8d88c