Fix foreign tabs with > 6 navigations not showing up on Android.
The code being deleted here was introduced in
http://crrev.com/23514039 because tabs with invalid indexes
were being received. I have now identified the cause of that
original issue, which is that the current index of a tab
isn't updated when we truncate the navigation stack for
syncing. For now, I'm making Android behave like all other
clients and bounding the index inside the navigations array.

BUG=378992

Review URL: https://codereview.chromium.org/433633002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@286879 4ff67af0-8c30-449e-8e8b-ad334ec8d88c