sync: Refactor app sync integration tests
Refactors the app sync integration tests so they no longer inspect
internal state or rely on self-notifications.  Introduces
AppsMatchChecker, which automatically installs synced apps while waiting
for a list of profiles to have a matching set of apps.  It's used to
implement AwaitAllProfilesHaveSameAppsAsVerifier().

Deletes two 'Disabled' tests for having little value.

I'm not sure what was wrong with the 'Merge' test (crbug.com/226055).
My hope is that it was related to the flakiness of AwaitQuiescence or
crbug.com/397766, since both of those are no longer be a problem in this
test.  I'm going to try re-enabling it and soon we'll see whether or not
it's actually fixed.

BUG=97780,95742,226055

Review URL: https://codereview.chromium.org/431133002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@287002 4ff67af0-8c30-449e-8e8b-ad334ec8d88c