Augment PrefHashFilter with the ability to cleanup deprecated tracked preferences and their matching MAC.
The time has come where some tracked preferences are starting to be deprecated and keeping them declared
solely for the sake of having them migrated back to the unprotected store by TrackedPreferencesMigrator
only to be cleared from that store shortly after is adding overhead (and furthermore with that technique
only the value would be cleaned up, not its matching MAC).

Implemented in CleanupDeprecatedTrackedPreferences() in pref_hash_filter.cc and added a comment pointing
to it from kTrackedPrefs.

Review URL: https://codereview.chromium.org/432533003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@287050 4ff67af0-8c30-449e-8e8b-ad334ec8d88c