Improve error checking in BookmarkModelAssociator::BuildAssociations.
We have crash reports where GetChromeNodeFromSyncId is returning NULL.
This may indicate we failed to create a BookmarkNode correspdoning to an
existing sync node.  Instead of crashing, return an unrecoverable
SyncError if GetChromeNodeFromSyncId returns NULL.

CreateBookmarkNode may fail (return NULL) if the supplied index is
invalid.  DCHECK that the index is valid and return an unrecoverable
SyncError if we fail to create the bookmark node.

BUG=398826

Review URL: https://codereview.chromium.org/432963002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@287062 4ff67af0-8c30-449e-8e8b-ad334ec8d88c