Remember user decisions on invalid certificates behind a flag
Currently if a user clicks thorough an invalid certificate, the decision is
remembered for the rest of the session, but after session restart, the decision
is forgotten. This implements a flag that remembers the decision for a specified
length of time across session restarts. Users can specific in the flag how long
decisions should be remembered, or they can specify that it should still be
reset on restart.

BUG=262615
NOTRY=true

Review URL: https://codereview.chromium.org/369703002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@287118 4ff67af0-8c30-449e-8e8b-ad334ec8d88c