Ensure public account policy wallpaper is stored in correct place.
Previously, when user_id_hash was empty, the wallpaper was stored with the wrong
filename and thus couldn't be retrieved later on.

Also fixes the problem that setting a wallpaper with small dimensions after a
wallpaper with large dimensions didn't clear all scaled copies of the previous
wallpaper.

BUG=388819

Review URL: https://codereview.chromium.org/422243002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@287313 4ff67af0-8c30-449e-8e8b-ad334ec8d88c