Translate: Bug fix: Wrong way to check if the window is minimized
Before this CL, to check if the window is minimized, !IsActive() was
used. This function returns false when the bubble window is focused on
Windows or Chrome OS. Therefore, the bubble was not updated when the
bubble is already shown.

This CL fixes this bug by replacing !IsActive() with IsMinimize() to
check the state of Window properly.

BUG=393136, 389014 (the original CL about the bug around minimized window)
TEST=manual

Review URL: https://codereview.chromium.org/428583004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@287448 4ff67af0-8c30-449e-8e8b-ad334ec8d88c