Allow recommended locales to be set for public sessions
This CL introduces the |SessionLocales| policy that can be used to
recommend one or more UI locales for a public session. If the policy is
set, the first recommended locale is pre-selected when starting a public
session. If more than one locale is recommended, the public session pod
is shown in its advanced form, highlighting the availability of language
and keyboard layout pickers.

The keyboard layout is automatically set to the most popular layout
matching the UI locale.

This CL also fixes the flaky DeviceLocalAccountTest.* and
TermsOfServiceDownloadTest.Ü.

BUG=214904, 241790, 401010
TEST=Extensive browser test coverage, including picker UI

Review URL: https://codereview.chromium.org/426063005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@287767 4ff67af0-8c30-449e-8e8b-ad334ec8d88c