Merge navigation list and directory tree.
This CL merges navigation list (the very left pane) and directory tree (middle pane).
In the conbined tree, the first-level children are VolumeItem or ShortcutItem, which correspond to NavigationModelItem.
Only Drive volume has children and the Drive volume. They are DirectoryItem and the behavior is similar to previous directory tree.

BUG=390848,397222
TEST=run browser_tests

Review URL: https://codereview.chromium.org/441233002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@287933 4ff67af0-8c30-449e-8e8b-ad334ec8d88c