Add a browser test for functions in notification provider API
Add a basic browser test for functions in notification provider API.
This is mainly to test the connection and provide a skeleton for future
tests. The functions are not fully implemented yet, and more tests will
be added as the functions are implemented.

BUG=

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=287692

Review URL: https://codereview.chromium.org/416423006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@287972 4ff67af0-8c30-449e-8e8b-ad334ec8d88c