Use PageManager in About page
This creates a PageManager.Observer class which delegates handling of
page history and title changes to an uber page observer:
- page_manager.js
- uber_page_manager_observer.js

The About page content is separated into its own file so in a future CL
it can be used as an overlay on another page.

The custom overlay implementation in help_base_page.js is no longer needed.

R=dbeam@chromium.org
BUG=313244
TBR=ben@chromium.org (for c/b/browser_resources.grd)

Review URL: https://codereview.chromium.org/423533003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@287989 4ff67af0-8c30-449e-8e8b-ad334ec8d88c