Reland: Parse Drive API responses all at once in the blocking pool.
Previous implementation did the parsing of string to base::Value
on blocking pool, and that of base::Value to specific data types
either on UI thread or on yet another post to blocking pool.

The previous implementation is slightly inefficient and moreover
involves a subtle bug 284244.

BUG=284244, 401843

This CL once landed as r288017 (Patch Set 4) and reverted.
This is the second try.

Review URL: https://codereview.chromium.org/442193002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288275 4ff67af0-8c30-449e-8e8b-ad334ec8d88c