Refactor SyncManager::Init to receive args via a struct.
SyncManager::Init is implemented in several places, takes a whole lot of
args, and in the near future is going to get a few more.

In some cases, many of the args are optional and it is tedious to pass a
bunch of NULLs or empty strings.  Use a struct instead of an arg list.

BUG=

Review URL: https://codereview.chromium.org/451743002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288280 4ff67af0-8c30-449e-8e8b-ad334ec8d88c