[Hotword] Do not open about:voicesearch page in incognito to prevent crash.
Voice search is not available in incognito mode so if 
prompted to open the about:voicesearch page while in an 
incognito window, instead redirect and open it in a new tab 
in the parent profile's current window. This is how other 
about: pages handle this situation.

BUG=400151
TEST=open an incognito window, navigate to about://voicesearch,
observe that it opens in a new tab in the parent window and does not crash.

Review URL: https://codereview.chromium.org/450763002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288289 4ff67af0-8c30-449e-8e8b-ad334ec8d88c