Adding the VerifyModule function (and helpers) to safe browsing.
This function allows us to compare a module in memory with 
its original on disk, and identify any differences (after 
accounting for relocations).  The state of the module 
(MODULE_UNKNOWN, MODULE_UNMODIFIED or MODULE_MODIFIED) is 
returned along with a list of the possibly modified exported 
functions of the module.


BUG=401854

Review URL: https://codereview.chromium.org/406043003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288319 4ff67af0-8c30-449e-8e8b-ad334ec8d88c