Change the title of the desktop User Manager widget.
The title of the window widget should be the product name (Chrome/Chromium).

Added a "close" button to make the last bubble in the tutorial dismissible.

BUG=399698,399697
TEST=Start Chrome with --new-avatar-menu. Choose "Switch person" from the
avatar menu of a brand new profile. Once you go through all the slides in
the tutorial, the last bubble which points to the "Add Person" button
should have a little x button in the top right corner. Also, the title of
the window should be "Chrome".

Review URL: https://codereview.chromium.org/440713005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288385 4ff67af0-8c30-449e-8e8b-ad334ec8d88c