Remember the configured device scale factor for external displays.
Users can specify the device scale factor for some external
displays. Now, it remembers the configured device scale factor
to the local storage, and recovers to the specified value when
reconnected.

This also fixes the bug when the user specifies the resolution
and the device scale factor at the same time (DSF was ignored
beforehand).

BUG=396704
R=oshima@chromium.org
TEST=manually, unit_tests

Review URL: https://codereview.chromium.org/455443002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288412 4ff67af0-8c30-449e-8e8b-ad334ec8d88c