[DevTools] Make DevTools clients talk directly to DevToolsAgentHost instead of using DevToolsManager.
Previously, client used methods on DevToolsManager passing |this| to most of them.
Instead, it should hold a scoped_refptr to inspected DevToolsAgentHost and talk
to it directly. This creates a single clean entry point for the client, and allows
in the future to handle browser-wide protocol messages and debug multiple hosts at once.

BUG=401073
TBR=kalman@chromium.org

Review URL: https://codereview.chromium.org/449043002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288452 4ff67af0-8c30-449e-8e8b-ad334ec8d88c