[PrintPreview] Custom input focus toggles
In print preview mode, focus toggles for radio buttons for pagerange
settings when custom input radio button is clicked again and page 
range is empty. 

This is not desired and happens because focus is set for
another radio button on blur event for custom input radio
button when it is empty.

BUG=397094

Review URL: https://codereview.chromium.org/435393005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288485 4ff67af0-8c30-449e-8e8b-ad334ec8d88c