Remove WeakHandle from components/invalidation
Replaces the use of WeakHandle with combinations of WeakPtr and
SingleThreadTaskRunner.  The WeakHandle encapsulates both of these
concepts, so the replacment is not much different from the original.

The removal makes the code a bit uglier.  However, the WeakHandle is not
widely used outside of sync and has no chance of making it to base/.
Its use in new code is discouraged.  Moving components/invalidation/ to
WeakPtr and SingeThreadTaskRunner will make it more future proof, and
allow us to remove its dependency on libsync.

TBR=dcheng
BUG=394925

Review URL: https://codereview.chromium.org/446223002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288523 4ff67af0-8c30-449e-8e8b-ad334ec8d88c