[Mac] Sign-in email of a profile without an error shouldn't look clickable.
In the new avatar bubble, we display the primary email. If this email
isn't in an authentication-error state, it isn't clickable, so it shouldn't
get a hand cursor when hovered over.

As a bonus, I've also cleaned up BuildLabel(), which was over-excited about
setting a background color, when all it wanted was to be transparent :)

BUG=401840
TEST=Start Chrome with --enable-new-avatar-menu (new flag change, hurrah!).
The profile's email address should be black, not clickable, and the mouse
cursor shouldn't change when hovering over it.

Review URL: https://codereview.chromium.org/454093002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288646 4ff67af0-8c30-449e-8e8b-ad334ec8d88c