Custom launcher pages: right-click menu is now the platform apps menu.
Added new extensions::ViewType, VIEW_TYPE_LAUNCHER_PAGE, which is used
to identify custom launcher page views and give them the platform app
menu. Note that currently some of the menu items have no effect.

Right-clicking on a custom launcher page that is not actually a platform
app results in a DCHECK failure (the --custom-launcher-page switch is a
use-at-your-own-risk switch; this situation will not be possible once we
switch to manifests).

BUG=399123

Review URL: https://codereview.chromium.org/442653002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288670 4ff67af0-8c30-449e-8e8b-ad334ec8d88c