Make SingleThreadProxy a SchedulerClient
This makes ui::Compositor no longer in charge of
scheduling commits and draws, deferring it to cc::Scheduler.

Other compositors that use SingleThreadProxy are left calling composite
synchronously and now pass a flag to indicate that this is their
intention.  This patch doesn't remove synchronous composite, but now
makes it mutually exclusive with scheduling.

BUG=329552, 287250

Review URL: https://codereview.chromium.org/134623005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288866 4ff67af0-8c30-449e-8e8b-ad334ec8d88c