VAVEA: Turn VAAPI encoder flag into a kill switch.
Flip the flag to enable VAVEA by default.

Also update the CastStreamingApiTest to not require VP8, as we may be using
H264 as well, but just assert on non-empty string for the codec. Do the
same for the audio codec while we are here.


TEST=encoding scenarios
BUG=378962,398322

Review URL: https://codereview.chromium.org/374553002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288920 4ff67af0-8c30-449e-8e8b-ad334ec8d88c