Update support for alphaEnabled (transparent windows) API
Rename transparentBackground -> alphaEnabled (per the spec).

Add app.window.alphaEnabled permission to control which apps are allowed to use alphaEnabled when creating windows.

Add browsertests to verify that the API is only available when permission had been granted and Beta channel or newer.

Add a whitelist (only used in Stable) for this API that defaults to the same set of apps as for app.window.shape.

BUG=260810

Review URL: https://codereview.chromium.org/190533006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@288983 4ff67af0-8c30-449e-8e8b-ad334ec8d88c