[Win] Profile desktop shortcuts should use the Gaia avatar if it's in use.
With --google-profile-info or --enable-new-avatar-menu turned on, signed in
profiles can use the account's Gaia image as the profile avatar. The
desktop shortcut shouldn't be surprised about this :)

Screenshot of the badged shortcuts in all their glory:
https://drive.google.com/folderview?id=0B1B1Up4p2NRMUjlTSU1EYUlBVG8&usp=sharing

BUG=402162
TEST=Flip --google-profile-info or --enable-new-avatar-menu. Sign in a profile,
and change the profile's avatar to the account's Gaia image. The desktop shortcut
should update and use the Gaia image correctly.

Review URL: https://codereview.chromium.org/462713002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@289100 4ff67af0-8c30-449e-8e8b-ad334ec8d88c