Implement support for <extensionoptions> embedding in WebUI
Modify permissions and checks in GuestView, WebUI, and
extensions to allow <extensionoptions> to be used within WebUI.
In the GuestView infrastructure, allow guest views in WebUI to
be created without extension ids. In extensions, allow extension
options-related APIs to be used in WebUI contexts. In WebUI,
relax the CSP to allow internal browser plugins.  


BUG=386842

Review URL: https://codereview.chromium.org/453613002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@289233 4ff67af0-8c30-449e-8e8b-ad334ec8d88c