Move app list search result icon sizes to app_list_constants.
This CL is a pure refactor of app list search result icon sizes which were
previously constants scattered about the codebase. This CL puts the icon
size as a static function in SearchResult so that there is only one place
the size needs to be changed.

BUG=391348

Review URL: https://codereview.chromium.org/458983004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@289517 4ff67af0-8c30-449e-8e8b-ad334ec8d88c