Route newly created notifications to notification provider API.
Whenever a new notification is created, route it to an
extension/app that has notification provider permission.
Currently, if there's an app with notificationProvider
permission, notifications will go to both message center
and that app. In the future, when additional sections in
Chrome Settings is implemented, this will send notifications
to only one party based on the user's selection.

BUG=

Review URL: https://codereview.chromium.org/441753002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@289532 4ff67af0-8c30-449e-8e8b-ad334ec8d88c