Integrate About page into Settings for Chrome OS settings in a window.
This is the last patch to get the About overlay showing in the Settings page
on Chrome OS when Settings in a window is enabled, so that we can hide the
uber navigation frame and truly display Settings in its own window.

BUG=379348
R=dbeam@chromium.org,ben@chromium.org

ben: please review:
    chrome/browser/browser_about_handler.cc
i hope i'm not doing anything blasphemous here!

Review URL: https://codereview.chromium.org/449623003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@289613 4ff67af0-8c30-449e-8e8b-ad334ec8d88c